naicsData json file is complete.
