--- Price_Pattern/WebContent/ED/bower.json
@@ -1,26 +1,0 @@
-{
-  "name": "EventDrops",
-  "version": "0.1.2",
-  "homepage": "https://github.com/marmelab/EventDrops",
-  "authors": [
-    "ThieryMichel <thiery@marmelab.com>"
-  ],
-  "description": "A time based / event series interactive visualization using d3.js.",
-  "main": "src/eventDrops.js",
-  "keywords": [
-    "d3",
-    "chart",
-    "event"
-  ],
-  "dependencies": {
-    "d3": "^3.4.11"
-  },
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ]
-}
