+++ Price_Pattern/WebContent/ED/example/custom/index.html
@@ -41,7 +41,7 @@
             
 
            arr = JSON.parse( arr); //
-         //  document.write(arr[0].Dates);
+         
             var data = [];
             var endTime = new Date("2015-12-31");
             var oneMonth = 30 * 24 * 60 * 60 * 1000;
