--- Price_Pattern/WebContent/ED/lib/filterData.js
@@ -1,19 +1,0 @@
-"use strict";
-/* global module */
-
-module.exports = function filterDate(data, scale) {
-  data = data || [];
-  var filteredData = [];
-  var boundary = scale.range();
-  var min = boundary[0];
-  var max = boundary[1];
-  data.forEach(function (datum) {
-    var value = scale(datum);
-    if (value < min || value > max) {
-      return;
-    }
-    filteredData.push(datum);
-  });
-
-  return filteredData;
-};
