--- Price_Pattern/WebContent/ED/lib/main.js
@@ -1,16 +1,0 @@
-"use strict";
-/* global require, define, module */
-
-var eventDrops = require('./eventDrops');
-
-if (typeof define === "function" && define.amd) {
-  define('d3.chart.eventDrops', ["d3"], function (d3) {
-    d3.chart = d3.chart || {};
-    d3.chart.eventDrops = eventDrops(d3);
-  });
-} else if (window) {
-  window.d3.chart = window.d3.chart || {};
-  window.d3.chart.eventDrops = eventDrops(window.d3);
-} else {
-  module.exports = eventDrops;
-}
