--- Price_Pattern/WebContent/ED/lib/util/configurable.js
@@ -1,16 +1,0 @@
-module.exports = function configurable(targetFunction, config, listeners) {
-  listeners = listeners || {};
-  for (var item in config) {
-    (function(item) {
-      targetFunction[item] = function(value) {
-        if (!arguments.length) return config[item];
-        config[item] = value;
-        if (listeners.hasOwnProperty(item)) {
-          listeners[item](value);
-        }
-
-        return targetFunction;
-      };
-    })(item); // for doesn't create a closure, forcing it
-  }
-};
